backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
stripe==7.4.0
bcrypt==4.0.1
PyJWT==2.8.0
pytest==7.4.2
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0 